esnet_gui_deployment/olderversions/web/src/MouseTrapBackendFiles/predict24values/requirements.txt
# Function dependencies, for example:
# package>=version


numpy
pandas
statsmodels
datetime 
simplejson
google-cloud-core
google-cloud-storage
google-cloud-firestore

